Help entries

Estas entradas pueden ser vistas en el REPL de Julia usando ? name.

FundamentosDMC.Options — Type

Structure containing the options and parameters of the simulation.

Examples

julia> opt = Options()

julia> opt = Options(nsteps=2000)

FundamentosDMC.Point2D — Type

Structure that contains a two-dimensional point, subtype of FieldVector from StaticArrays.

FundamentosDMC.Point3D — Type

Structure that contains a three-dimensional point, subtype of FieldVector from StaticArrays.

FundamentosDMC.System — Type

Structure containing the system initial coordinates and box sides.

Example

julia> sys = System(n=100,sides=[100,100])

FundamentosDMC.forces! — Method

forces!(x::Vector{T},f::Vector{T},opt::Options) where T

Subroutine that computes the force. It modifies the input f vector.

FundamentosDMC.image — Method

image(x,sides)

Move coordinates to minimum perioddic image.

FundamentosDMC.kinetic — Method

Function that computes the kinetic energy.

FundamentosDMC.mc — Method

mc(sys::System{T},opt=Options=Options()) where T

Peforms a Monte-Carlos simulation.

FundamentosDMC.md — Method

md(sys,opt) where T

Performs a simple MD simulation.

FundamentosDMC.md_berendsen — Method

md_berendsen(sys::System{T},opt::Options) where T

Performs a MD simulation with a Berendsen bath.

FundamentosDMC.md_isokinetic — Method

md_isokinetic(sys::System{T},opt::Options) where T

Performs a MD simulation with an isokinetic bath.

FundamentosDMC.md_langevin — Method

md_langevin(sys::System{T},opt::Options) where T

Performs a MD simulation with a Langevin bath.

FundamentosDMC.minimize! — Method

minimize!(sys::System{T},opt::Options)

Minimizes the potential energy of sys.x0. Modifies sys.x0.

Example

julia> sys = System(n=100,sides=[100,100])

julia> minimize!(sys,Options(tol=1e-3))
Energy before minimization: 38322.72337856496
Energy after minimization: -74.15646912098042

FundamentosDMC.potential — Method

Function that computes the potential energy

FundamentosDMC.printxyz — Method

printxyz(time,x,sys,buffer)

Subroutine that writes the current coordinates to an output file, previously opened as buffer.

FundamentosDMC.printxyz — Method

printxyz(x,sys,file::String) 

Prints one set of coordinates, given in x to file file. Overwrites the file.

FundamentosDMC.radial_distribution — Method

radial_distribution(trajfile::String,dmax::Float64=10.,nbins::Int=10,T::DataType=Point2D)

Computes the g(r). If T=Point3D the system will be read as three-dimensional.

FundamentosDMC.velocities — Method

velocities(sys::System{T},opt::Options) where T

Initialize velocities